storage: Move SSTableInfo* structs to sst_info.go
Given these structs are common across storage engines,
it made little sense to keep these in rocksdb.go.

Release note: None.
